Move household config (suggestions, categories, recipe tags) out of nav drawer
Suggestions, suggestion categories, and recipe tags are household-level
configuration, not global admin navigation items. They now live under
/households/[id]/ with domain grouping (Lists / Recipes) to support
future extensibility of each area.

Changes:
- Remove Suggestions, Suggestion Categories, Recipe Tags from AppLayout drawer
- Add Configuration section to household detail page, grouped by domain:
  - Lists: Suggestions → /households/[id]/lists/suggestions
           Suggestion Categories → /households/[id]/lists/suggestions/categories
  - Recipes: Recipe Tags → /households/[id]/recipes/tags
- Configuration section is admin-only (same access control as before)
- Old /admin/suggestions* and /admin/recipes/tags routes redirect to /households
- Move test coverage to new locations; update household detail tests
- Update visual regression tests to new page routes
