+++ anything-frontend/src/app/admin/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
 
 export default function AdminPage() {
-  redirect("/admin/suggestions");
+  redirect("/households");
 }
